Started Building AVL Tree
Began building the AVL Tree, by starting with the rotation functions.
Extended some of the AVL Node functionality to account for determining
which leaf of a parent a node is on. Tested changes with JUNIT and
extended JUNIT tests for changes. Next step is to build JUNIT test for t
he rotation functions. Then Build and test the double rotation
functions.