Added Balancer Class
I built a class consisting of static functions that balance the nodes. I
did this at the advice of stack overflow suggesting that if I wanted to
run Unit tests on private functions than they should be broken out into
a separate. I decided I really needed those extra unit tests for the
balancing functions so I created a new class that would perform all of
the balancing operations.

While testing the balancing I came across several bugs in the AVL node
that had to do with accessing NULL children. So I extended the unit
tests for the AVL Node to include tests on NULL children and I added
error checking for NULL children in the AVL Node.