Built the Insertion Function
Built the main insertion function with node balancer routines. Have not
written the JUNIT tests for the insertion function. Did some very basic
testing of the implementation in the main file to verify it's working.
Also built the prinTree, isLessThan, isGreaterThen, and areEquals
routines. Notable bugs/Oversights,  that were found and fixed: when the
root node is balanced there is a new root node, when doing
lesser/greater/equal comparisons we need to check for null values.