Expanded Delete Functions & AVLTree Insert Tests
Initial changes consisted of building more complete JUNIT tests for the
insert object routine of the avl tree. I also redefined the avl node's
cleanNode routine as a delete node routine and expanded it to remove the
node from the parent. I also built the deleteTree and printTree routines
in the AVL tree. Finally I started building the findNode routine in the
AVL Tree when I was interrupted by Peepers.